requirements.txt
dash
flask
kafka-python
numpy
pandas
